pyproject.toml
[tool.poetry]
name = "trading-bot"
version = "0.1.0"
description = ""
authors = ["gchu <gchu@zendesk.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pandas = ">=2.1.4"
requests = "^2.31.0"
jupyterlab = "^4.2.0"
pytz = "^2024.1"
seaborn = "^0.13.2"
huggingface-hub = {extras = ["cli"], version = "^0.23.0"}
numpy = ">=1.26.4"
utilsforecast = "^0.1.10"
python-dotenv = "^1.0.1"
yfinance = "^0.2.40"
pandera = "^0.20.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.10.1"
ruff = "^0.5.1"
pandas-stubs = "^2.2.2.240603"
types-requests = "^2.32.0.20240712"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
